Webclient for ExchangeRateApiV4

+++ src/main/java/com/jackosky/currency/dto/ConversionRequest.java
@@ -1,5 +1,6 @@
 package com.jackosky.currency.dto;
 
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
@@ -7,5 +8,5 @@ public class ConversionRequest {
 
   private final String from;
   private final String to;
-  private final double amount;
+  private final BigDecimal amount;
 }
