+++ src/main/java/com/jackosky/currency/dto/ConversionResponse.java
@@ -1,5 +1,6 @@
 package com.jackosky.currency.dto;
 
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
@@ -7,6 +8,6 @@ public class ConversionResponse {
 
   private final String from;
   private final String to;
-  private final double amount;
-  private final double converted;
+  private final BigDecimal amount;
+  private final BigDecimal converted;
 }
